+++ Main/src/main/java/com/main/acad/error/ConnectionPoolFailedException.java
@@ -1,0 +1,7 @@
+package com.main.acad.error;
+
+public class ConnectionPoolFailedException extends RuntimeException {
+    public ConnectionPoolFailedException(String message) {
+        super(message);
+    }
+}
