+++ Main/src/main/java/com/main/acad/service/ChaptersServiceImplementation.java
@@ -1,17 +1,18 @@
 package com.main.acad.service;
 
-import com.main.acad.dao.ChapterDaoImplementation;
+import com.main.acad.dao.ChapterDao;
+import com.main.acad.dao.SimpleChapterDao;
 import com.main.acad.entity.Chapter;
+import com.main.acad.error.ChapterDaoFailedExeption;
 
 import java.io.FileReader;
 import java.io.IOException;
-import java.sql.Connection;
 import java.util.List;
 import java.util.logging.Logger;
 
 public class ChaptersServiceImplementation implements ChaptersService {
 
-    private ChapterDaoImplementation dao = ChapterDaoImplementation.getInstance();
+    private ChapterDao dao = SimpleChapterDao.getInstance();
     private static ChaptersServiceImplementation instance;
     private static final Logger logger = Logger.getLogger(ChaptersServiceImplementation.class.getName());
 
@@ -26,7 +27,7 @@ public class ChaptersServiceImplementation implements ChaptersService {
     }
 
     @Override
-    public String getInformstioAboutChildren(String name) {
+    public String getInformstioAboutChildren(String name) throws ChapterDaoFailedExeption{
         FileReader fileReader = this.dao.getInformstioAboutChildren(name);
         StringBuilder filePage = new StringBuilder();
         int count;
@@ -36,18 +37,18 @@ public class ChaptersServiceImplementation implements ChaptersService {
             }
             logger.info("File : " + name + " successfully read");
         } catch (IOException e) {
-            logger.info("An error occurred in the StrategySimulationServiceImplementation class in the getInformstioAboutChildren method");
+            logger.info("An error occurred in the ChaptersServiceImplementation class in the getInformstioAboutChildren method");
         }
         return filePage.toString();
     }
 
     @Override
-    public List<Chapter> listChapters() {
+    public List<Chapter> listChapters()throws ChapterDaoFailedExeption {
         return dao.getlistChapters();
     }
 
     @Override
-    public List<Chapter> listChildren(int id) {
+    public List<Chapter> listChildren(int id)throws ChapterDaoFailedExeption {
         return dao.getlistChildren(id);
     }
