+++ Main/src/main/java/com/main/acad/service/UsersService.java
@@ -1,0 +1,4 @@
+package com.main.acad.service;
+
+public interface UsersService {
+}
