+++ Main/src/main/webapp/WEB-INF/web.xml
@@ -25,35 +25,35 @@
         <url-pattern>/api/*</url-pattern>
     </servlet-mapping>
 
-    <servlet>
-        <servlet-name>LoginServlet</servlet-name>
-        <servlet-class>com.main.acad.servlet.LoginServlet</servlet-class>
-    </servlet>
+    <!--<servlet>-->
+        <!--<servlet-name>LoginServlet</servlet-name>-->
+        <!--<servlet-class>com.main.acad.servlet.LoginServlet</servlet-class>-->
+    <!--</servlet>-->
 
-    <servlet-mapping>
-        <servlet-name>LoginServlet</servlet-name>
-        <url-pattern>/login</url-pattern>
-    </servlet-mapping>
-    
-    <servlet>
-        <servlet-name>RegistrationServlet</servlet-name>
-        <servlet-class>com.main.acad.servlet.RegistrationServlet</servlet-class>
-    </servlet>
-    
-    <servlet-mapping>
-        <servlet-name>RegistrationServlet</servlet-name>
-        <url-pattern>/registration</url-pattern>
-    </servlet-mapping>
+    <!--<servlet-mapping>-->
+        <!--<servlet-name>LoginServlet</servlet-name>-->
+        <!--<url-pattern>/login</url-pattern>-->
+    <!--</servlet-mapping>-->
+    <!---->
+    <!--<servlet>-->
+        <!--<servlet-name>RegistrationServlet</servlet-name>-->
+        <!--<servlet-class>com.main.acad.servlet.RegistrationServlet</servlet-class>-->
+    <!--</servlet>-->
+    <!---->
+    <!--<servlet-mapping>-->
+        <!--<servlet-name>RegistrationServlet</servlet-name>-->
+        <!--<url-pattern>/registration</url-pattern>-->
+    <!--</servlet-mapping>-->
 
-    <servlet>
-        <servlet-name>CreateNewUserServlet</servlet-name>
-        <servlet-class>com.main.acad.servlet.CreateNewUserServlet</servlet-class>
-    </servlet>
+    <!--<servlet>-->
+        <!--<servlet-name>CreateNewUserServlet</servlet-name>-->
+        <!--<servlet-class>com.main.acad.servlet.CreateNewUserServlet</servlet-class>-->
+    <!--</servlet>-->
 
-    <servlet-mapping>
-        <servlet-name>CreateNewUserServlet</servlet-name>
-        <url-pattern>/createNewUser</url-pattern>
-    </servlet-mapping>
+    <!--<servlet-mapping>-->
+        <!--<servlet-name>CreateNewUserServlet</servlet-name>-->
+        <!--<url-pattern>/createNewUser</url-pattern>-->
+    <!--</servlet-mapping>-->
 
     <filter>
         <filter-name>HttpServletFilter</filter-name>
