+++ Main/src/main/webapp/registration.html
@@ -137,7 +137,7 @@
 </head>
 <body>
 <script src="http://code.jquery.com/jquery-latest.min.js"></script>
-<script type="text/javascript" src="login.js" charset="UTF-8"></script>
+<script type="text/javascript" src="registration.js" charset="UTF-8"></script>
 <div id="page">
     <header id="header">
         <div id="header-inner">
